leapscholar-brand-monitor/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
textblob>=0.17.0
vaderSentiment>=3.3.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0 